lib/storage: move legacy code to legacy files
Follow-up for f97f627 (#8134)
