+++ lib/storage/storage.go
@@ -6,7 +6,6 @@ import (
 	"math"
 	"os"
 	"path/filepath"
-	"regexp"
 	"sort"
 	"sync"
 	"sync/atomic"
@@ -709,11 +708,6 @@ func (s *Storage) UpdateMetrics(m *Metrics) {
 	}
 }
 
-// TODO(@rtm0): Move to storage_legacy.go
-func (s *Storage) legacyNextRetentionSeconds() int64 {
-	return s.legacyNextRotationTimestamp.Load() - int64(fasttime.UnixTimestamp())
-}
-
 // SetFreeDiskSpaceLimit sets the minimum free disk space size of current storage path
 //
 // The function must be called before opening or creating any storage.
@@ -772,34 +766,6 @@ func (s *Storage) startFreeDiskSpaceWatcher() {
 func (s *Storage) notifyReadWriteMode() {
 	s.tb.NotifyReadWriteMode()
 	s.legacyNotifyReadWriteMode()
-}
-
-// TODO(@rtm0): Move to storage_legacy.go
-func (s *Storage) startLegacyRetentionWatcher() {
-	if !s.hasLegacyIndexDBs() {
-		return
-	}
-	s.legacyRetentionWatcherWG.Add(1)
-	go func() {
-		s.legacyRetentionWatcher()
-		s.legacyRetentionWatcherWG.Done()
-	}()
-}
-
-// TODO(@rtm0): Move to storage_legacy.go
-func (s *Storage) legacyRetentionWatcher() {
-	for {
-		d := s.legacyNextRetentionSeconds()
-		select {
-		case <-s.stopCh:
-			return
-		case currentTime := <-time.After(time.Second * time.Duration(d)):
-			s.legacyMustRotateIndexDB(currentTime)
-			if !s.hasLegacyIndexDBs() {
-				return
-			}
-		}
-	}
 }
 
 func (s *Storage) startCurrHourMetricIDsUpdater() {
@@ -1095,42 +1061,6 @@ func (s *Storage) mustSaveCache(c *workingsetcache.Cache, name string) {
 
 // saveCacheLock prevents from data races when multiple concurrent goroutines save the same cache.
 var saveCacheLock sync.Mutex
-
-// LegacySetRetentionTimezoneOffset sets the offset, which is used for
-// calculating the time for legacy indexdb rotation.
-//
-// See https://github.com/VictoriaMetrics/VictoriaMetrics/pull/2574
-//
-// TODO(@rtm0): Move to storage_legacy.go
-func LegacySetRetentionTimezoneOffset(offset time.Duration) {
-	legacyRetentionTimezoneOffsetSecs = int64(offset.Seconds())
-}
-
-// TODO(@rtm0): Move to storage_legacy.go
-var legacyRetentionTimezoneOffsetSecs int64
-
-// TODO(@rtm0): Move to storage_legacy.go
-func legacyNextRetentionDeadlineSeconds(atSecs, retentionSecs, offsetSecs int64) int64 {
-	// Round retentionSecs to days. This guarantees that per-day inverted index works as expected
-	const secsPerDay = 24 * 3600
-	retentionSecs = ((retentionSecs + secsPerDay - 1) / secsPerDay) * secsPerDay
-
-	// Schedule the deadline to +4 hours from the next retention period start
-	// because of historical reasons - see https://github.com/VictoriaMetrics/VictoriaMetrics/issues/248
-	offsetSecs -= 4 * 3600
-
-	// Make sure that offsetSecs doesn't exceed retentionSecs
-	offsetSecs %= retentionSecs
-
-	// align the retention deadline to multiples of retentionSecs
-	// This makes the deadline independent of atSecs.
-	deadline := ((atSecs + offsetSecs + retentionSecs - 1) / retentionSecs) * retentionSecs
-
-	// Apply the provided offsetSecs
-	deadline -= offsetSecs
-
-	return deadline
-}
 
 func (s *Storage) getMetricNameFromCache(dst []byte, metricID uint64) []byte {
 	// There is no need in checking for deleted metricIDs here, since they
@@ -2533,83 +2463,6 @@ func (s *Storage) getTSIDFromCache(dst *legacyTSID, metricName []byte) bool {
 func (s *Storage) storeTSIDToCache(tsid *legacyTSID, metricName []byte) {
 	buf := (*[unsafe.Sizeof(*tsid)]byte)(unsafe.Pointer(tsid))[:]
 	s.tsidCache.Set(metricName, buf)
-}
-
-// TODO(@rtm0): Move to storage_legacy.go
-func (s *Storage) mustOpenLegacyIndexDBTables(path string) *legacyIndexDBs {
-	if !fs.IsPathExist(path) {
-		return nil
-	}
-
-	// Search for the two most recent tables: prev and curr.
-
-	// Placing the regexp inside the func in order to keep legacy code close to
-	// each other and because this function is called only once on startup.
-	indexDBTableNameRegexp := regexp.MustCompile("^[0-9A-F]{16}$")
-	des := fs.MustReadDir(path)
-	var tableNames []string
-	for _, de := range des {
-		if !fs.IsDirOrSymlink(de) {
-			// Skip non-directories.
-			continue
-		}
-		tableName := de.Name()
-		if !indexDBTableNameRegexp.MatchString(tableName) {
-			// Skip invalid directories.
-			continue
-		}
-		tableDirPath := filepath.Join(path, tableName)
-		if fs.IsPartiallyRemovedDir(tableDirPath) {
-			// Finish the removal of partially deleted directory, which can occur
-			// when the directory was removed during unclean shutdown.
-			fs.MustRemoveDir(tableDirPath)
-			continue
-		}
-		tableNames = append(tableNames, tableName)
-	}
-	sort.Slice(tableNames, func(i, j int) bool {
-		return tableNames[i] < tableNames[j]
-	})
-
-	if len(tableNames) > 3 {
-		// Remove all the tables except the last three tables.
-		for _, tn := range tableNames[:len(tableNames)-3] {
-			pathToRemove := filepath.Join(path, tn)
-			logger.Infof("removing obsolete indexdb dir %q...", pathToRemove)
-			fs.MustRemoveDir(pathToRemove)
-			logger.Infof("removed obsolete indexdb dir %q", pathToRemove)
-		}
-		fs.MustSyncPath(path)
-		tableNames = tableNames[len(tableNames)-3:]
-	}
-	if len(tableNames) == 3 {
-		// Also remove next idb.
-		pathToRemove := filepath.Join(path, tableNames[2])
-		logger.Infof("removing next indexdb dir %q...", pathToRemove)
-		fs.MustRemoveDir(pathToRemove)
-		logger.Infof("removed next indexdb dir %q", pathToRemove)
-		fs.MustSyncPath(path)
-		tableNames = tableNames[:2]
-	}
-
-	numIDBs := len(tableNames)
-	legacyIDBs := &legacyIndexDBs{}
-
-	if numIDBs == 0 {
-		return nil
-	}
-
-	if numIDBs > 1 {
-		currPath := filepath.Join(path, tableNames[1])
-		legacyIDBs.idbCurr = mustOpenLegacyIndexDB(currPath, s)
-	}
-
-	if numIDBs > 0 {
-		prevPath := filepath.Join(path, tableNames[0])
-		legacyIDBs.idbPrev = mustOpenLegacyIndexDB(prevPath, s)
-	}
-
-	return legacyIDBs
 }
 
 // wasMetricIDMissingBefore checks if passed metricID was already registered as missing before.
